+++ bench/tx-generator/src/Cardano/Benchmarking/Script/Core.hs
@@ -30,10 +30,10 @@ import           Cardano.Benchmarking.LogTypes as Core (AsyncBenchmarkControl (.
 import           Cardano.Benchmarking.OuroborosImports as Core (LocalSubmitTx, SigningKeyFile,
                    makeLocalConnectInfo, protocolToCodecConfig)
 import           Cardano.Benchmarking.Script.Aeson (prettyPrintOrdered, readProtocolParametersFile)
-import qualified Cardano.Benchmarking.Script.Ogmios as OgmiosBackend
-import qualified Cardano.Benchmarking.Script.Submission as Submission
 import           Cardano.Benchmarking.Script.Env hiding (Error (TxGenError))
 import qualified Cardano.Benchmarking.Script.Env as Env (Error (TxGenError))
+import qualified Cardano.Benchmarking.Script.Ogmios as OgmiosBackend
+import qualified Cardano.Benchmarking.Script.Submission as Submission
 import           Cardano.Benchmarking.Script.Types
 import           Cardano.Benchmarking.Types as Core (SubmissionErrorPolicy (..))
 import           Cardano.Benchmarking.Version as Version
@@ -246,11 +246,12 @@ submitInEra submitMode generator txParams era = do
     NodeToNode _ -> error "NodeToNode deprecated: ToDo: remove"
     Benchmark nodes tpsRate txCount -> benchmarkTxStream txStream nodes tpsRate txCount era
     LocalSocket -> submitAll (void . localSubmitTx . Utils.mkTxInModeCardano) txStream
-    SubmitToEndpoint endpointType (EndpointUri uri) -> case endpointType of
-      Ogmios -> Submission.runSubmitTransport
-                  (Submission.onRejectionFor generator)
-                  (OgmiosBackend.withOgmiosTransport uri)
-                  txStream
+    SubmitToEndpoint endpoint -> case endpoint of
+      OgmiosEndpoint (EndpointUri uri) ->
+        Submission.runSubmitTransport
+          (Submission.onRejectionFor generator)
+          (OgmiosBackend.withOgmiosTransport uri)
+          txStream
     DumpToFile filePath -> liftIO $ Streaming.writeFile filePath $ Streaming.map showTx txStream
     DiscardTX -> liftIO $ Streaming.mapM_ forceTx txStream
  where