+++ bench/tx-generator/src/Cardano/Benchmarking/Script/Types.hs
@@ -33,6 +33,7 @@ module Cardano.Benchmarking.Script.Types (
         , ProtocolParametersSource(QueryLocalNode, UseLocalProtocolFile)
         , ScriptBudget(AutoScript, StaticScriptBudget)
         , ScriptSpec(..)
+        , SubmissionEndpoint(..)
         , SubmissionEndpointProtocol(..)
         , SubmitMode(Benchmark, DiscardTX, DumpToFile, LocalSocket,
                 NodeToNode, SubmitToEndpoint)
@@ -47,7 +48,7 @@ import           Cardano.Benchmarking.OuroborosImports (SigningKeyFile)
 import           Cardano.Node.Configuration.NodeAddress (NodeIPv4Address)
 import           Cardano.TxGenerator.ProtocolParameters (ProtocolParameters)
 import           Cardano.TxGenerator.Setup.NixService (EndpointUri (..), NodeDescription,
-                   SubmissionEndpointProtocol (..))
+                   SubmissionEndpoint (..), SubmissionEndpointProtocol (..))
 import           Cardano.TxGenerator.Types
 
 import           Prelude
@@ -188,9 +189,10 @@ data SubmitMode where
   DumpToFile  :: !FilePath -> SubmitMode
   DiscardTX   :: SubmitMode
   NodeToNode  :: NonEmpty NodeIPv4Address -> SubmitMode --deprecated
-  -- | Submit through an external service at the given URI, with the
-  -- 'SubmissionEndpointProtocol' selecting which backend protocol to speak.
-  SubmitToEndpoint :: !SubmissionEndpointProtocol -> !EndpointUri -> SubmitMode
+  -- | Submit through an external service, with the 'SubmissionEndpoint'
+  -- carrying both the backend protocol to speak and the endpoint it
+  -- addresses.
+  SubmitToEndpoint :: !SubmissionEndpoint -> SubmitMode
   deriving (Show, Eq)
 deriving instance Generic SubmitMode